--- src/Main.java
@@ -1,10 +1,0 @@
-import manage.*;
-import resources.*;
-
-// здесь только тестировал
-public class Main {
-    public static void main(String[] args) {
-
-// заменил папку
-    }
-}
