--- src/manage/HistoryManager.java
@@ -1,13 +1,0 @@
-package manage;
-
-import resources.Task;
-
-import java.util.List;
-
-public interface HistoryManager {
-    void add(Task task);
-
-    void remove(int id);
-
-    List<Task> getHistory();
-}
