--- src/manage/Manager.java
@@ -1,13 +1,0 @@
-package manage;
-
-
-    public class Manager {
-        public static InMemoryTaskManager getDefault() {
-            return new InMemoryTaskManager();
-        }
-
-        public static InMemoryHistoryManager getDefaultHistory() {
-            return new InMemoryHistoryManager();
-        }
-    }
-
