+++ src/model/Status.java
@@ -1,0 +1,7 @@
+package model;
+
+public enum Status {
+    NEW,
+    IN_PROGRESS,
+    DONE
+}
