--- src/resources/Epic.java
@@ -1,33 +1,0 @@
-package resources;
-
-import java.util.ArrayList;
-
-public class Epic extends Task {
-
-    private ArrayList<Subtask> subtaskList = new ArrayList<>();
-
-    public Epic(int id, String name, String description, Status status) {
-        super(id, name, description, status);
-    }
-
-    public Epic(String name, String description) {
-        super(name, description);
-    }
-
-    public void addSubtask(Subtask subtask) {
-        subtaskList.add(subtask);
-    }
-
-    public void clearSubtasks() {
-        subtaskList.clear();
-    }
-
-
-    public ArrayList<Subtask> getSubtaskList() {
-        return subtaskList;
-    }
-
-    public void setSubtaskList(ArrayList<Subtask> subtaskList) {
-        this.subtaskList = subtaskList;
-    }
-}
