--- src/resources/Status.java
@@ -1,7 +1,0 @@
-package resources;
-
-public enum Status {
-    NEW,
-    IN_PROGRESS,
-    DONE
-}
