--- src/resources/Subtask.java
@@ -1,26 +1,0 @@
-package resources;
-
-public class Subtask extends Task {
-    private final int epicID;
-
-    public Subtask(int id, String name, String description, Status status, int epicID) {
-        super(id, name, description, status);
-        this.epicID = epicID;
-    }
-
-    public Subtask(String name, String description, int epicID) {
-        super(name, description);
-        this.epicID = epicID;
-    }
-
-    public int getEpicID() {
-        return epicID;
-    }
-
-    @Override
-    public String toString() {
-        return "Subtask{" +
-                "epicID=" + epicID +
-                "} " + super.toString();
-    }
-}
