+++ src/service/HistoryManager.java
@@ -1,0 +1,13 @@
+package service;
+import model.Task;
+
+import java.util.List;
+
+public interface HistoryManager {
+
+    void add(Task task);
+
+    void remove(int id);
+
+    List<Task> getHistory();
+}
