+++ src/service/Managers.java
@@ -1,0 +1,11 @@
+package service;
+
+public class Managers {
+    public static InMemoryTaskManager getDefault() {
+        return new InMemoryTaskManager();
+    }
+
+    public static InMemoryHistoryManager getDefaultHistory() {
+        return new InMemoryHistoryManager();
+    }
+}
