+++ test/model/EpicTest.java
@@ -1,16 +1,12 @@
-package test.resources;
-
+package test.model;
+import model.*;
 import org.junit.jupiter.api.Test;
-import resources.Epic;
-import resources.Status;
+import static org.junit.jupiter.api.Assertions.*;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-
-class EpicTest {
+public class EpicTest {
 
     @Test
-    public void epicsWithEqualIdShouldBeEqual() {
+    public void EpicsWithEqualIdShouldBeEqual() {
         Epic epic1 = new Epic(10, "Сделать ремонт", "Уложиться в 2 миллиона", Status.NEW);
         Epic epic2 = new Epic(10, "Подготовиться к собеседованию", "1 июля в 11:00",
                 Status.IN_PROGRESS);
