+++ test/service/InMemoryHistoryManagerTest.java
@@ -1,27 +1,22 @@
-package manage;
-
+package test.service;
+import model.*;
+import service.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import resources.Epic;
-import resources.Status;
-import resources.Subtask;
-import resources.Task;
-
+import static org.junit.jupiter.api.Assertions.*;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 public class InMemoryHistoryManagerTest {
-
 
     private static TaskManager taskManager;
     private static InMemoryHistoryManager historyManager;
 
     @BeforeEach
     public void beforeEach() {
-        historyManager = Manager.getDefaultHistory();
-        taskManager = Manager.getDefault();
+        historyManager = Managers.getDefaultHistory();
+        taskManager = Managers.getDefault();
     }
+
 
     @Test
     public void getHistoryShouldReturnOldTaskAfterUpdate() {
@@ -96,7 +91,7 @@ public class InMemoryHistoryManagerTest {
     }
 
     @Test
-    void testRemoveAllTask() {
+    void testRemoveAllTask(){
         Task washFloor = new Task("Написать проект", "Сдать в срок");
         Task goToTeach = new Task("пойти в школу", "Учить информатику");
         Epic flatRenovation = new Epic("Новая задача", "Проверить");
@@ -149,4 +144,5 @@ public class InMemoryHistoryManagerTest {
         assertTrue(history.contains(task3));
         assertFalse(history.contains(task2)); // task2 должен быть удален
     }
+
 }
