+++ test/service/ManagersTest.java
@@ -1,0 +1,22 @@
+package test.service;
+import model.*;
+
+import org.junit.jupiter.api.Test;
+import service.InMemoryHistoryManager;
+import service.InMemoryTaskManager;
+import service.Managers;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ManagersTest {
+
+    @Test
+    void getDefaultShouldInitializeInMemoryTaskManager() {
+        assertInstanceOf(InMemoryTaskManager.class, Managers.getDefault());
+    }
+
+    @Test
+    void getDefaultHistoryShouldInitializeInMemoryHistoryManager() {
+        assertInstanceOf(InMemoryHistoryManager.class, Managers.getDefaultHistory());
+    }
+}
